test: Added tests for Linux file URIs
