+++ src/Uno.UI.Runtime.Skia.Gtk/ApplicationModel/DataTransfer/ClipboardExtensions.cs
@@ -224,7 +224,7 @@ namespace Uno.UI.Runtime.Skia.GTK.Extensions.ApplicationModel.DataTransfer
 					foreach (var item in items)
 					{
 						var path = item.Path;
-						builder.AppendLine(UrlEncode(path));
+						builder.AppendLine(FileUriHelper.UrlEncode(path));
 					}
 
 					nativeData.Set(GnomeCopiedFilesContent, 8, Encoding.UTF8.GetBytes(builder.ToString().Trim()));
@@ -298,39 +298,6 @@ namespace Uno.UI.Runtime.Skia.GTK.Extensions.ApplicationModel.DataTransfer
 		private void Clipboard_OwnerChange(object o, OwnerChangeArgs args)
 		{
 			ContentChanged?.Invoke(this, args);
-		}
-
-		/// <summary>
-		/// Encodes a file path to a file:// Url.
-		/// While the built-in Uri class can handle this, it does not process files
-		/// such as '/home/user/%51.txt' correctly.
-		/// </summary>
-		/// <param name="path">Path to the file</param>
-		/// <returns>file:// url to the file</returns>
-		private string UrlEncode(string path)
-		{
-			var uri = new StringBuilder();
-			foreach (var ch in path)
-			{
-				if (('a' <= ch && ch <= 'z') || ('A' <= ch && ch <= 'Z') || ('0' <= ch && ch <= '9') ||
-					"-._~".Contains(ch))
-				{
-					uri.Append(ch);
-				}
-				else if (ch == Path.DirectorySeparatorChar || ch == Path.AltDirectorySeparatorChar)
-				{
-					uri.Append('/');
-				}
-				else
-				{
-					var bytes = Encoding.UTF8.GetBytes(new[] { ch });
-					foreach (var b in bytes)
-					{
-						uri.Append($"%{b:X2}");
-					}
-				}
-			}
-			return "file://" + uri;
 		}
 	}
 }
